+++ eventservice/api/controllers/events.py
@@ -362,6 +362,39 @@ def patch(event_id):
         __logger.exception(ex)
         abort(405)
 
+    group_memberships = list()
+    try:
+        _, group_memberships = get_group_memberships()
+    except Exception as ex:
+        __logger.exception(ex)
+        abort(500)
+
+    db = None
+    event = None
+    try:
+        db = get_db()
+        event = db['events'].find_one({'_id': ObjectId(event_id)}, {'_id': 0})
+    except Exception as ex:
+        __logger.exception(ex)
+        abort(500)
+
+    if not group_memberships:
+        # check public group
+        if event and event.get('isGroupPrivate') is True:
+            abort(401)
+    else:
+        # get event and check the group id
+        if event:
+            found = False
+            for group_member in group_memberships:
+                if event.get('createdByGroupId') == group_member.get('id'):
+                    if group_member.get('role') == 'admin':
+                        found = True
+                        break
+            if not found:
+                abort(401)
+
+
     try:
         db = get_db()
         status = db['events'].update_one({'_id': ObjectId(event_id)}, {"$set": req_data})